upgrade to ScalaTest 3.1; update build to use sbt-tpolecat; ScalaJS 0.6.31

+++ core/src/main/scala/com/dwolla/testutils/random/Fakes.scala
@@ -26,7 +26,7 @@ trait Fakes {
 
   def double(): Double = rnd.nextDouble()
 
-  def float(): Double = rnd.nextFloat()
+  def float(): Double = rnd.nextFloat().toDouble
 
   def int(n: Int): Int = rnd.nextInt(n)
